refactor(basket): Phase 3 layout move Infrastructure Data/Redis
- Move CustomerBasketCacheModel, InMemoryBasketStore into Infrastructure/Data/Redis/.
- Namespace becomes Basket.Service.Infrastructure.Data.Redis; no behavior change.
