+++ basket-microservice/Basket.Service/Infrastructure/Data/Redis/CustomerBasketCacheModel.cs
@@ -1,5 +1,5 @@
 using Basket.Service.Domain;
 
-namespace Basket.Service.Infrastructure.Data;
+namespace Basket.Service.Infrastructure.Data.Redis;
 
 internal record CustomerBasketCacheModel(List<BasketProduct> Products);
